Reverting changes to latch, keeping refactorings of counting_semaphore and sliding_semaphore
