+++ libs/core/synchronization/include/hpx/synchronization/detail/counting_semaphore.hpp
@@ -29,24 +29,25 @@ namespace hpx { namespace lcos { namespace local { namespace detail {
         typedef lcos::local::spinlock mutex_type;
 
     public:
-        HPX_EXPORT counting_semaphore(std::ptrdiff_t value = 0);
-        HPX_EXPORT ~counting_semaphore();
+        HPX_CORE_EXPORT counting_semaphore(std::ptrdiff_t value = 0);
+        HPX_CORE_EXPORT ~counting_semaphore();
 
-        HPX_EXPORT void wait(
+        HPX_CORE_EXPORT void wait(
             std::unique_lock<mutex_type>& l, std::ptrdiff_t count);
 
-        HPX_EXPORT bool wait_until(std::unique_lock<mutex_type>& l,
+        HPX_CORE_EXPORT bool wait_until(std::unique_lock<mutex_type>& l,
             util::steady_time_point const& abs_time, std::ptrdiff_t count);
 
-        HPX_EXPORT bool try_wait(
+        HPX_CORE_EXPORT bool try_wait(
             std::unique_lock<mutex_type>& l, std::ptrdiff_t count = 1);
 
-        HPX_EXPORT bool try_acquire(std::unique_lock<mutex_type>& l);
+        HPX_CORE_EXPORT bool try_acquire(std::unique_lock<mutex_type>& l);
 
-        HPX_EXPORT void signal(
+        HPX_CORE_EXPORT void signal(
             std::unique_lock<mutex_type> l, std::ptrdiff_t count);
 
-        HPX_EXPORT std::ptrdiff_t signal_all(std::unique_lock<mutex_type> l);
+        HPX_CORE_EXPORT std::ptrdiff_t signal_all(
+            std::unique_lock<mutex_type> l);
 
     private:
         std::ptrdiff_t value_;
