+++ libs/core/synchronization/include/hpx/synchronization/mutex.hpp
@@ -20,9 +20,11 @@ namespace hpx { namespace threads {
 
     using thread_id_type = thread_id;
     using thread_self = coroutines::detail::coroutine_self;
+
     /// The function \a get_self_id returns the HPX thread id of the current
     /// thread (or zero if the current thread is not a HPX thread).
     HPX_CORE_EXPORT thread_id_type get_self_id();
+
     /// The function \a get_self_ptr returns a pointer to the (OS thread
     /// specific) self reference to the current HPX thread.
     HPX_CORE_EXPORT thread_self* get_self_ptr();
@@ -65,7 +67,7 @@ namespace hpx { namespace lcos { namespace local {
     protected:
         mutable mutex_type mtx_;
         threads::thread_id_type owner_id_;
-        detail::condition_variable cond_;
+        lcos::local::detail::condition_variable cond_;
     };
 
     ///////////////////////////////////////////////////////////////////////////
