+++ libs/core/synchronization/src/detail/counting_semaphore.cpp
